feat(#13): fix all the qulice suggestions
