+++ src/main/java/org/eolang/opeo/ast/Invocation.java
@@ -1,0 +1,96 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016-2023 Objectionary.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package org.eolang.opeo.ast;
+
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+/**
+ * Invocation output node.
+ * @since 0.1
+ */
+public final class Invocation implements AstNode {
+
+    /**
+     * Source or target on which the invocation is performed.
+     */
+    private final AstNode source;
+
+    /**
+     * Method name.
+     */
+    private final String method;
+
+    /**
+     * Arguments.
+     */
+    private final List<AstNode> arguments;
+
+    /**
+     * Constructor.
+     * @param source Source or target on which the invocation is performed
+     * @param method Method name
+     * @param arguments Arguments
+     */
+    public Invocation(
+        final AstNode source,
+        final String method,
+        final List<AstNode> arguments
+    ) {
+        this.source = source;
+        this.method = method;
+        this.arguments = arguments;
+    }
+
+    @Override
+    public String print() {
+        return String.format(
+            "(%s).%s%s",
+            this.source.print(),
+            this.method,
+            this.args()
+        );
+    }
+
+    @Override
+    public String identifier() {
+        return UUID.randomUUID().toString();
+    }
+
+    /**
+     * Print arguments.
+     * @return Arguments
+     */
+    private String args() {
+        final String result;
+        if (this.arguments.isEmpty()) {
+            result = "";
+        } else {
+            result = this.arguments.stream().map(AstNode::print)
+                .collect(Collectors.joining(" ", " ", ""));
+        }
+        return result;
+    }
+}
